requirements.txt
Django>=5.2,<5.3
Pillow>=10.0
django-allauth[socialaccount]>=65.0,<66
